Debris reentry data

A subset of data from the following catalogues * satcat Standard Satellite Catalog (S) Objects from the US Satellite Catalog * auxcat Auxiliary Satellite Catalog (A) Objects omitted from the US catalog * tmpcat Temporary Catalog (T) Objects expected to receive an S catalog number soon consisting of only reentries occurred in the 25 years from 2000-01-01 (included) till 2025-01-01 (excluded) `satcat`, `auxcat` and `tmpcat` from J. C. McDowell, “General Catalog of Artificial Space Objects.” 2020.

Usage

reentries_gcat

Format

## `reentries_gcat` A data frame with 1968 rows and 6 columns: * `jcat`: A space object tag; a single prefix letter followed by a sequence number with 5 digits. Possible prefixes are: - `A` for auxcat (Auxiliary catalog) - `S` for stdcat (Standard catalog) - `T` for tmpcat (Temporary catalog) * `piece`: The piece designation for the object * `type`: This field tells you something about what kind of object: - `R1`–`R5`: Stage number for launch vehicle stage. - `C..[A|M]`: a component, i.e. a functional part of the spacecraft or rocket that is deliberately or accidentally ejected. * `inclination`: the orbit inclination in the [-90, 90] interval. * `reentry_date`: the date of reentry in ["vague date" format](https://planet4589.org/space/gcat/web/intro/vague.html). * `dry_mass`: the (poosibly estimated) dry mass.

Source

<https://planet4589.org/space/gcat>